backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for the farm watch portal"
authors = ["jeafreezy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.18.3"
fastapi = "^0.85.0"
pydantic = "^1.10.2"
black = "^22.8.0"
SQLAlchemy = "^1.4.41"
GeoAlchemy2 = "^0.12.5"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"
SQLAlchemy-Utils = "^0.38.3"
alembic = "^1.8.1"
python-multipart = "^0.0.5"



[tool.poetry.group.dev.dependencies]
mypy = "^0.981"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
